requirements.txt
pynonymizer
python-dotenv
